+++ src/lib.rs
@@ -20,6 +20,8 @@ pub mod rocket;
 
 // Forward declarations
 pub use credentials::Credentials;
+pub use sessions::user::Session as UserSession;
+pub use sessions::service_account::Session as ServiceSession;
 
 /// Authentication trait.
 ///